requirements.txt
torch
tables
pandas
dill
tensorboardX
cox
python-git
gym
mujoco-py<2.1,>=2.0